Refactor parallel loop and fix compiler warnings
Restructure OpenMP parallelization in Integral.cpp using collapse(2)
instead of ordered sections for better scalability. Move file I/O
outside the parallel region and pre-compute grid values.

Additional changes:
- Suppress unused parameter warnings with (void) casts in Process.hpp,
  GSLVEGAS.cpp, and YM.cpp
- Fix signed/unsigned comparison in constants.cpp power_recursive()
- Update YM.cpp to use configurable output file path with colored output
- Update plot.gp to read from YM.dat instead of result.dat

+++ src/constants.cpp
@@ -24,7 +24,7 @@ constexpr T power_recursive(T base) {
 
     // Recursive step: x^n = x * x^(n-1)
     T result = 1;
-    for (int i = 0; i < exp; ++i) {
+    for (size_t i = 0; i < exp; ++i) {
         result *= base;
     }
     return result;
